requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
wandb>=0.12.0
pyyaml>=6.0
tqdm>=4.62.0
joblib>=1.1.0
